+++ lib/generators/active_bot/templates/bot_template.rb
@@ -1,0 +1,3 @@
+class <%= class_name %>Bot < ActiveBot::Base
+  # logic for your bot
+end