+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## [0.1.0] - 2024-01-14
 
 - Initial release
+- Add `ActiveBot::Bot` class
+- Add generator for `ActiveBot::Bot` class